Preserve 64-bit last insert rowids across triggers. (CVS 2779)
FossilOrigin-Name: 4dd7cfaa587fd5f76a0b60db0f02d5f57f801564

+++ src/vdbeInt.h
@@ -250,7 +250,7 @@ struct Fifo {
 */
 typedef struct Context Context;
 struct Context {
-  int lastRowid;    /* Last insert rowid (sqlite3.lastRowid) */
+  i64 lastRowid;    /* Last insert rowid (sqlite3.lastRowid) */
   int nChange;      /* Statement changes (Vdbe.nChanges)     */
   Fifo sFifo;       /* Records that will participate in a DELETE or UPDATE */
 };
